testing principle: widgets should probably not interface directly with objects, that's what the communicationmanager should be for?
